add fastcache
